Learned that Australian Patience with one redraw has an official name, Canberra. Replace all instances of Australian Patience (One Redraw) with Canberra. No new code.
