+++ app/src/main/java/com/heyzeusv/solitaire/util/Enums.kt
@@ -43,7 +43,7 @@ enum class Games(
     KLONDIKE_TURN_ONE(R.string.games_klondike_turn_one, 1, Game.GAME_KLONDIKETURNONE),
     KLONDIKE_TURN_THREE(R.string.games_klondike_turn_three, 3, Game.GAME_KLONDIKETURNTHREE),
     AUSTRALIAN_PATIENCE(R.string.games_australian_patience, 1, Game.GAME_AUSTRALIAN_PATIENCE),
-    AUSTRALIAN_PATIENCE_ONE_REDRAW(R.string.games_australian_patience_one_redraw, 1, Game.GAME_AUSTRALIAN_PATIENCE_ONE_REDRAW)
+    CANBERRA(R.string.games_canberra, 1, Game.GAME_CANBERRA)
 }
 
 /**
