+++ app/src/main/proto/stat_prefs.proto
@@ -23,5 +23,5 @@ enum Game {
   GAME_KLONDIKETURNONE = 0;
   GAME_KLONDIKETURNTHREE = 1;
   GAME_AUSTRALIAN_PATIENCE = 2;
-  GAME_AUSTRALIAN_PATIENCE_ONE_REDRAW = 3;
+  GAME_CANBERRA = 3;
 }
