+++ app/src/main/res/values/strings.xml
@@ -34,7 +34,7 @@
     <string name="games_klondike_turn_one">Klondike (Turn One)</string>
     <string name="games_klondike_turn_three">Klondike (Turn Three)</string>
     <string name="games_australian_patience">Australian Patience</string>
-    <string name="games_australian_patience_one_redraw">Australian Patience (One Redraw)</string>
+    <string name="games_canberra">Canberra</string>
     <string name="suits_clubs">Clubs</string>
     <string name="suits_diamonds">Diamonds</string>
     <string name="suits_hearts">Hearts</string>
